+++ etc/config.xml
@@ -24,10 +24,10 @@
             </calculator_config>
         </svea_part_payment_calculator>
         <payment>
-            <svea_facelift_payment>
+            <svea_paynowpaylatersplit_payment>
                 <active>0</active>
-                <title>Svea facelift Payment</title>
-                <model>SveaFaceliftPaymentFacade</model>
+                <title>Svea Pay Now / Pay Later Split</title>
+                <model>SveaPayNowPayLaterSplitPaymentFacade</model>
                 <order_status>pending</order_status>
                 <is_preselect_required>1</is_preselect_required>
                 <payment_action>authorize</payment_action>
@@ -46,7 +46,7 @@
                 <payment_method_subgroup_2_method_filter>FI54,PIVO,SIIR,APPL,GPAY</payment_method_subgroup_2_method_filter>
                 <payment_method_subgroup_1_title>Online Banking</payment_method_subgroup_1_title>
                 <payment_method_subgroup_1_method_filter>FI01,FI02,FI03,FI04,FI05,FI06,FI07,FI08,FI09,FI10,FI11,FI12,FI13,FI14,FI15</payment_method_subgroup_1_method_filter>
-            </svea_facelift_payment>
+            </svea_paynowpaylatersplit_payment>
             <svea_collated_payment>
                 <active>0</active>
                 <title>Svea Collated Payment</title>
